Move enums and models into separate folders
