+++ libs/importer/src/importers/lastpass/access/enums/duo-factor.ts
@@ -1,0 +1,6 @@
+export enum DuoFactor {
+  Push,
+  Call,
+  Passcode,
+  SendPasscodesBySms,
+}
