+++ libs/importer/src/importers/lastpass/access/enums/duo-status.ts
@@ -1,0 +1,5 @@
+export enum DuoStatus {
+  Success,
+  Error,
+  Info,
+}
