+++ libs/importer/src/importers/lastpass/access/enums/idp-provider.ts
@@ -1,0 +1,8 @@
+export enum IdpProvider {
+  Azure = 0,
+  OktaAuthServer = 1,
+  OktaNoAuthServer = 2,
+  Google = 3,
+  PingOne = 4,
+  OneLogin = 5,
+}
