+++ libs/importer/src/importers/lastpass/access/enums/lastpass-login-type.ts
@@ -1,0 +1,5 @@
+export enum LastpassLoginType {
+  MasterPassword = 0,
+  // Not sure what Types 1 and 2 are?
+  Federated = 3,
+}
