+++ libs/importer/src/importers/lastpass/access/models/index.ts
@@ -1,0 +1,10 @@
+export { Account } from "./account";
+export { Chunk } from "./chunk";
+export { ClientInfo } from "./client-info";
+export { FederatedUserContext } from "./federated-user-context";
+export { OobResult } from "./oob-result";
+export { OtpResult } from "./otp-result";
+export { ParserOptions } from "./parser-options";
+export { Session } from "./session";
+export { SharedFolder } from "./shared-folder";
+export { UserTypeContext } from "./user-type-context";
