+++ libs/importer/src/importers/lastpass/access/models/user-type-context.ts
@@ -1,17 +1,19 @@
+import { IdpProvider, LastpassLoginType } from "../enums";
+
 export class UserTypeContext {
-  type: Type;
+  type: LastpassLoginType;
   IdentityProviderGUID: string;
   IdentityProviderURL: string;
   OpenIDConnectAuthority: string;
   OpenIDConnectClientId: string;
   CompanyId: number;
-  Provider: Provider;
+  Provider: IdpProvider;
   PkceEnabled: boolean;
   IsPasswordlessEnabled: boolean;
 
   isFederated(): boolean {
     return (
-      this.type === Type.Federated &&
+      this.type === LastpassLoginType.Federated &&
       this.hasValue(this.IdentityProviderURL) &&
       this.hasValue(this.OpenIDConnectAuthority) &&
       this.hasValue(this.OpenIDConnectClientId)
@@ -21,19 +23,4 @@ export class UserTypeContext {
   private hasValue(str: string) {
     return str != null && str.trim() !== "";
   }
-}
-
-export enum Provider {
-  Azure = 0,
-  OktaAuthServer = 1,
-  OktaNoAuthServer = 2,
-  Google = 3,
-  PingOne = 4,
-  OneLogin = 5,
-}
-
-export enum Type {
-  MasterPassword = 0,
-  // Not sure what Types 1 and 2 are?
-  Federated = 3,
 }
