+++ libs/importer/src/importers/lastpass/access/parser.ts
@@ -1,12 +1,9 @@
 import { CryptoFunctionService } from "@bitwarden/common/platform/abstractions/crypto-function.service";
 import { Utils } from "@bitwarden/common/platform/misc/utils";
 
-import { Account } from "./account";
 import { BinaryReader } from "./binary-reader";
-import { Chunk } from "./chunk";
 import { CryptoUtils } from "./crypto-utils";
-import { ParserOptions } from "./parser-options";
-import { SharedFolder } from "./shared-folder";
+import { Account, Chunk, ParserOptions, SharedFolder } from "./models";
 
 const AllowedSecureNoteTypes = new Set<string>([
   "Server",
