+++ libs/importer/src/importers/lastpass/access/ui.ts
@@ -1,6 +1,5 @@
 import { DuoUi } from "./duo-ui";
-import { OobResult } from "./oob-result";
-import { OtpResult } from "./otp-result";
+import { OobResult, OtpResult } from "./models";
 
 export abstract class Ui extends DuoUi {
   // To cancel return OtpResult.Cancel, otherwise only valid data is expected.
